requirements.txt
pip install;
python-dotenv; 
fastapi "uvicorn[standard]";
pydantic;
langchain;
langchain-community; 
langchain-huggingface; 
langchain-google-genai; 
faiss-cpu;
pypdf;
langchain-text-splitters; 
sentence-transformers;
streamlit requests
